python/requirements.txt
flask
flask_cors
gunicorn
numpy
pandas
scikit-learn
tensorflow
keras
google-cloud-secret-manager
google-cloud-storage